Security: Authorization Checks for All Controllers

+++ lib/Controller/PolicyController.php
@@ -5,21 +5,37 @@ declare(strict_types=1);
 namespace OCA\PTO\Controller;
 
 use OCA\PTO\AppInfo\Application;
+use OCA\PTO\Service\AuthorizationService;
 use OCA\PTO\Service\PolicyService;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http;
 use OCP\AppFramework\Http\DataResponse;
 use OCP\IRequest;
+use OCP\IUserSession;
 
 class PolicyController extends Controller {
     private PolicyService $service;
+    private AuthorizationService $authService;
+    private IUserSession $userSession;
 
-    public function __construct(IRequest $request, PolicyService $service) {
+    public function __construct(
+        IRequest $request,
+        PolicyService $service,
+        AuthorizationService $authService,
+        IUserSession $userSession
+    ) {
         parent::__construct(Application::APP_ID, $request);
         $this->service = $service;
+        $this->authService = $authService;
+        $this->userSession = $userSession;
+    }
+
+    private function getUserId(): string {
+        return $this->userSession->getUser()->getUID();
     }
 
     /**
+     * List all policies (any authenticated user can view)
      * @NoAdminRequired
      * @NoCSRFRequired
      */
@@ -29,6 +45,7 @@ class PolicyController extends Controller {
     }
 
     /**
+     * Show a single policy (any authenticated user can view)
      * @NoAdminRequired
      * @NoCSRFRequired
      */
@@ -42,11 +59,15 @@ class PolicyController extends Controller {
     }
 
     /**
+     * Create a policy - admin only
      * @NoAdminRequired
      */
     public function create(): DataResponse {
         try {
-            // TODO: Check admin permission
+            if (!$this->authService->isAdmin($this->getUserId())) {
+                return new DataResponse(['error' => 'Only administrators can create policies'], Http::STATUS_FORBIDDEN);
+            }
+
             $data = json_decode(file_get_contents('php://input'), true);
             
             $policy = $this->service->create(
@@ -66,11 +87,15 @@ class PolicyController extends Controller {
     }
 
     /**
+     * Update a policy - admin only
      * @NoAdminRequired
      */
     public function update(int $id): DataResponse {
         try {
-            // TODO: Check admin permission
+            if (!$this->authService->isAdmin($this->getUserId())) {
+                return new DataResponse(['error' => 'Only administrators can update policies'], Http::STATUS_FORBIDDEN);
+            }
+
             $data = json_decode(file_get_contents('php://input'), true);
             
             $policy = $this->service->update(
@@ -92,11 +117,15 @@ class PolicyController extends Controller {
     }
 
     /**
+     * Delete a policy - admin only
      * @NoAdminRequired
      */
     public function destroy(int $id): DataResponse {
         try {
-            // TODO: Check admin permission
+            if (!$this->authService->isAdmin($this->getUserId())) {
+                return new DataResponse(['error' => 'Only administrators can delete policies'], Http::STATUS_FORBIDDEN);
+            }
+
             $this->service->delete($id);
             return new DataResponse([], Http::STATUS_NO_CONTENT);
         } catch (\Exception $e) {
